Move some stuff out of BellmanFord
To be used by Dijkstra also.
